requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
pyYAML

-e .